pyproject.toml
[tool.poetry]
name = "weight"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "0.115.0"
SQLAlchemy = "2.0.35"
uvicorn = "0.31.0"
psycopg2 = "2.9.9"
python-decouple = "3.8"
bcrypt = "4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
